servico_impressao/requirements.txt
python-escpos>=3.0
pyusb>=1.0
pywin32>=305; sys_platform == "win32"
